moves the tests near the implementation for the changed files
